backend/requirements.txt
asgiref==3.3.4
autopep8==1.7.0
boto3==1.24.79
botocore==1.27.79
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
dj-rest-auth==2.1.4
Django==3.2
django-allauth==0.44.0
django-cors-headers==3.13.0
django-filter==22.1
django-seed==0.3.1
django-storages==1.13.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-yasg==1.21.3
Faker==14.2.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
mysqlclient==2.0.3
Naked==0.1.31
oauthlib==3.1.0
packaging==21.3
Pillow==9.2.0
psycopg2==2.9.3
pycodestyle==2.9.1
pycparser==2.20
pycryptodome==3.15.0
pycryptodomex==3.15.0
PyJWT==2.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
toposort==1.7
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.4
